refactor: create TaskQueueBase to factorize code
